Revert "wip: major refactor"
This reverts commit 1ea3f373e886c7884239bbede96d81c45fa87771.
