--- src/core/context.cpp
@@ -1,412 +1,0 @@
-#include "context.hpp"
-
-#include <stdexcept>
-#include <iostream>
-#include <set>
-
-using namespace molten;
-
-static VkResult CreateDebugUtilsMessengerEXT(VkInstance instance,
-                                             const VkDebugUtilsMessengerCreateInfoEXT* pCreateInfo,
-                                             const VkAllocationCallbacks* pAllocator,
-                                             VkDebugUtilsMessengerEXT* pDebugMessenger)
-{
-    auto func = (PFN_vkCreateDebugUtilsMessengerEXT)vkGetInstanceProcAddr(
-        instance, "vkCreateDebugUtilsMessengerEXT");
-    if (func != nullptr)
-    {
-        return func(instance, pCreateInfo, pAllocator, pDebugMessenger);
-    }
-    else
-    {
-        return VK_ERROR_EXTENSION_NOT_PRESENT;
-    }
-}
-
-static void DestroyDebugUtilsMessengerEXT(VkInstance instance,
-                                          VkDebugUtilsMessengerEXT debugMessenger,
-                                          const VkAllocationCallbacks* pAllocator)
-{
-    auto func = (PFN_vkDestroyDebugUtilsMessengerEXT)vkGetInstanceProcAddr(
-        instance, "vkDestroyDebugUtilsMessengerEXT");
-    if (func != nullptr)
-    {
-        func(instance, debugMessenger, pAllocator);
-    }
-}
-
-void Context::init()
-{
-    create_instance();
-    setup_debug_messenger();
-    create_surface();
-    pick_physical_device();
-    create_logical_device();
-    create_command_pool();
-}
-
-void Context::init_window()
-{
-    glfwInit();
-    glfwWindowHint(GLFW_CLIENT_API, GLFW_NO_API);
-    glfwWindowHint(GLFW_RESIZABLE, GLFW_TRUE);
-
-    m_window = glfwCreateWindow(WIN_WIDTH, WIN_HEIGHT, "Molten", nullptr, nullptr);
-    glfwSetWindowUserPointer(m_window, this);
-    glfwSetFramebufferSizeCallback(m_window, framebuffer_resize_callback);
-}
-
-void Context::cleanup()
-{
-    vkDestroyCommandPool(m_device, m_command_pool, nullptr);
-
-    vkDestroyDevice(m_device, nullptr);
-
-    if (ENABLE_VALIDATION_LAYERS)
-        DestroyDebugUtilsMessengerEXT(m_instance, m_debug_messenger, nullptr);
-
-    vkDestroySurfaceKHR(m_instance, m_surface, nullptr);
-    vkDestroyInstance(m_instance, nullptr);
-
-    glfwDestroyWindow(m_window);
-    glfwTerminate();
-}
-
-void Context::framebuffer_resize_callback(GLFWwindow* window, int width, int height)
-{
-    auto context = reinterpret_cast<Context*>(glfwGetWindowUserPointer(window));
-    context->m_framebuffer_resized = true;
-}
-
-void Context::create_instance()
-{
-    printf("Creating Vulkan Instance - IN PROGRESS \n");
-
-    if (ENABLE_VALIDATION_LAYERS && !check_validation_layer_support())
-        throw std::runtime_error("Validation Layers Requested, But Not Available!");
-
-    // Create Vulkan Instance
-    VkApplicationInfo app_info = {};
-    app_info.sType = VK_STRUCTURE_TYPE_APPLICATION_INFO;
-    app_info.pApplicationName = "Molten";
-    app_info.applicationVersion = VK_MAKE_VERSION(1, 0, 0);
-    app_info.pEngineName = "Molten Engine";
-    app_info.engineVersion = VK_MAKE_VERSION(1, 0, 0);
-    app_info.apiVersion = VK_API_VERSION_1_0;
-
-    auto extensions = get_required_extensions();
-    VkDebugUtilsMessengerCreateInfoEXT debug_create_info = {};
-
-    VkInstanceCreateInfo create_info = {};
-    create_info.sType = VK_STRUCTURE_TYPE_INSTANCE_CREATE_INFO;
-    create_info.pApplicationInfo = &app_info;
-    create_info.enabledExtensionCount = static_cast<uint32_t>(extensions.size());
-    create_info.ppEnabledExtensionNames = extensions.data();
-    if (ENABLE_VALIDATION_LAYERS)
-    {
-        create_info.enabledLayerCount = static_cast<uint32_t>(VALIDATION_LAYERS.size());
-        create_info.ppEnabledLayerNames = VALIDATION_LAYERS.data();
-
-        populate_debug_messenger_create_info(debug_create_info);
-        create_info.pNext = (VkDebugUtilsMessengerCreateInfoEXT*)&debug_create_info;
-    }
-    else
-    {
-        create_info.enabledLayerCount = 0;
-        create_info.pNext = nullptr;
-    }
-
-    if (vkCreateInstance(&create_info, nullptr, &m_instance) != VK_SUCCESS)
-        throw std::runtime_error("failed to create instance!");
-
-    // Check for Extensions
-    uint32_t extension_count = 0;
-    vkEnumerateInstanceExtensionProperties(nullptr, &extension_count, nullptr);
-    std::vector<VkExtensionProperties> extensions_list(extension_count);
-    vkEnumerateInstanceExtensionProperties(nullptr, &extension_count, extensions_list.data());
-
-    std::cout << "Available extensions:\n";
-    for (const auto& extension : extensions_list)
-    {
-        std::cout << '\t' << extension.extensionName << '\n';
-    }
-
-    // Load Functions for Dynamic Rendering
-    BeginRendering =
-        (PFN_vkCmdBeginRenderingKHR)vkGetInstanceProcAddr(m_instance, "vkCmdBeginRenderingKHR");
-    EndRendering =
-        (PFN_vkCmdEndRenderingKHR)vkGetInstanceProcAddr(m_instance, "vkCmdEndRenderingKHR");
-
-    printf("\n Creating Vulkan Instance - DONE \n");
-}
-
-bool Context::check_validation_layer_support()
-{
-    uint32_t layer_count = 0;
-    vkEnumerateInstanceLayerProperties(&layer_count, nullptr);
-
-    std::vector<VkLayerProperties> available_layers(layer_count);
-    vkEnumerateInstanceLayerProperties(&layer_count, available_layers.data());
-
-    for (const char* layer_name : VALIDATION_LAYERS)
-    {
-        bool layer_found = false;
-
-        for (const auto& layer_properties : available_layers)
-        {
-            if (strcmp(layer_name, layer_properties.layerName) == 0)
-            {
-                layer_found = true;
-                break;
-            }
-        }
-
-        if (!layer_found)
-        {
-            return false;
-        }
-    }
-
-    return true;
-}
-
-std::vector<const char*> Context::get_required_extensions()
-{
-    uint32_t glfw_ext_count = 0;
-    const char** glfw_ext;
-    glfw_ext = glfwGetRequiredInstanceExtensions(&glfw_ext_count);
-
-    std::vector<const char*> extensions(glfw_ext, glfw_ext + glfw_ext_count);
-
-    if (ENABLE_VALIDATION_LAYERS)
-        extensions.push_back(VK_EXT_DEBUG_UTILS_EXTENSION_NAME);
-
-    return extensions;
-}
-
-VKAPI_ATTR VkBool32 VKAPI_CALL Context::debug_callback(
-    VkDebugUtilsMessageSeverityFlagBitsEXT message_severity,
-    VkDebugUtilsMessageTypeFlagsEXT message_type,
-    const VkDebugUtilsMessengerCallbackDataEXT* callback_data, void* user_data)
-{
-    std::cerr << "validation Layer: " << callback_data->pMessage << std::endl;
-
-    return VK_FALSE;
-}
-
-void Context::setup_debug_messenger()
-{
-    if (!ENABLE_VALIDATION_LAYERS)
-        return;
-
-    VkDebugUtilsMessengerCreateInfoEXT create_info{};
-    populate_debug_messenger_create_info(create_info);
-
-    if (CreateDebugUtilsMessengerEXT(m_instance, &create_info, nullptr, &m_debug_messenger) !=
-        VK_SUCCESS)
-    {
-        throw std::runtime_error("Failed to Set Up Debug Messenger!");
-    }
-}
-
-void Context::populate_debug_messenger_create_info(VkDebugUtilsMessengerCreateInfoEXT& create_info)
-{
-    create_info = {};
-    create_info = {};
-    create_info.sType = VK_STRUCTURE_TYPE_DEBUG_UTILS_MESSENGER_CREATE_INFO_EXT;
-    create_info.messageSeverity = VK_DEBUG_UTILS_MESSAGE_SEVERITY_VERBOSE_BIT_EXT |
-                                  VK_DEBUG_UTILS_MESSAGE_SEVERITY_WARNING_BIT_EXT |
-                                  VK_DEBUG_UTILS_MESSAGE_SEVERITY_ERROR_BIT_EXT;
-    create_info.messageType = VK_DEBUG_UTILS_MESSAGE_TYPE_GENERAL_BIT_EXT |
-                              VK_DEBUG_UTILS_MESSAGE_TYPE_VALIDATION_BIT_EXT |
-                              VK_DEBUG_UTILS_MESSAGE_TYPE_PERFORMANCE_BIT_EXT;
-    create_info.pfnUserCallback = debug_callback;
-}
-
-void Context::create_surface()
-{
-    if (glfwCreateWindowSurface(m_instance, m_window, nullptr, &m_surface) != VK_SUCCESS)
-        throw std::runtime_error("Failed To Create Window Surface!");
-}
-
-void Context::pick_physical_device()
-{
-    uint32_t device_count = 0;
-    vkEnumeratePhysicalDevices(m_instance, &device_count, nullptr);
-    if (device_count == 0)
-        throw std::runtime_error("Failed To Find GPUs With Vulkan Support!");
-
-    std::vector<VkPhysicalDevice> devices(device_count);
-    vkEnumeratePhysicalDevices(m_instance, &device_count, devices.data());
-
-    for (const auto& device : devices)
-    {
-        if (is_device_suitable(device))
-        {
-            m_physical_device = device;
-            break;
-        }
-    }
-
-    if (m_physical_device == VK_NULL_HANDLE)
-        throw std::runtime_error("Failed To Find a Suitable GPU!");
-}
-
-bool Context::is_device_suitable(VkPhysicalDevice device)
-{
-    QueueFamilyIndices indices = QueueFamilyIndices::find_queue_families(device, m_surface);
-
-    bool extensions_supported = check_device_extension_support(device);
-    bool swap_chain_adequate = false;
-    if (extensions_supported)
-    {
-        SwapChainSupportDetails swap_chain_support =
-            SwapChainSupportDetails::query_swap_chain_support(device, m_surface);
-        swap_chain_adequate =
-            !swap_chain_support.formats.empty() && !swap_chain_support.present_modes.empty();
-    }
-
-    return indices.is_complete() && extensions_supported && swap_chain_adequate;
-}
-
-void Context::create_logical_device()
-{
-    QueueFamilyIndices indices = QueueFamilyIndices::find_queue_families(m_physical_device, m_surface);
-
-    std::vector<VkDeviceQueueCreateInfo> queue_create_infos;
-    std::set<uint32_t> unique_queue_families = {indices.graphics_family.value(),
-                                                indices.present_family.value()};
-
-    float queue_priority = 1.0f;
-    for (uint32_t queue_family : unique_queue_families)
-    {
-        VkDeviceQueueCreateInfo queue_create_info = {};
-        queue_create_info.sType = VK_STRUCTURE_TYPE_DEVICE_QUEUE_CREATE_INFO;
-        queue_create_info.queueFamilyIndex = queue_family;
-        queue_create_info.queueCount = 1;
-        queue_create_info.pQueuePriorities = &queue_priority;
-        queue_create_infos.push_back(queue_create_info);
-    }
-
-    VkPhysicalDeviceFeatures device_features = {};
-
-    VkPhysicalDeviceDynamicRenderingFeaturesKHR dynamic_rendering_feature = {};
-    dynamic_rendering_feature.sType =
-        VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_DYNAMIC_RENDERING_FEATURES_KHR;
-    dynamic_rendering_feature.dynamicRendering = VK_TRUE;
-
-    VkDeviceCreateInfo create_info = {};
-    create_info.sType = VK_STRUCTURE_TYPE_DEVICE_CREATE_INFO;
-    create_info.pNext = &dynamic_rendering_feature;
-    create_info.queueCreateInfoCount = static_cast<uint32_t>(queue_create_infos.size());
-    create_info.pQueueCreateInfos = queue_create_infos.data();
-    create_info.pEnabledFeatures = &device_features;
-    create_info.enabledExtensionCount = static_cast<uint32_t>(DEVICE_EXTENSIONS.size());
-    create_info.ppEnabledExtensionNames = DEVICE_EXTENSIONS.data();
-    if (ENABLE_VALIDATION_LAYERS)
-    {
-        create_info.enabledLayerCount = static_cast<uint32_t>(VALIDATION_LAYERS.size());
-        create_info.ppEnabledLayerNames = VALIDATION_LAYERS.data();
-    }
-    else
-        create_info.enabledLayerCount = 0;
-
-    if (vkCreateDevice(m_physical_device, &create_info, nullptr, &m_device) != VK_SUCCESS)
-        throw std::runtime_error("Failed To Create Logical Device!");
-
-    vkGetDeviceQueue(m_device, indices.graphics_family.value(), 0, &m_graphics_queue);
-    vkGetDeviceQueue(m_device, indices.present_family.value(), 0, &m_present_queue);
-}
-
-bool Context::check_device_extension_support(VkPhysicalDevice device)
-{
-    uint32_t extension_count = 0;
-    vkEnumerateDeviceExtensionProperties(device, nullptr, &extension_count, nullptr);
-    std::vector<VkExtensionProperties> available_extensions(extension_count);
-    vkEnumerateDeviceExtensionProperties(device, nullptr, &extension_count,
-                                         available_extensions.data());
-
-    std::set<std::string> required_extensions(DEVICE_EXTENSIONS.begin(), DEVICE_EXTENSIONS.end());
-
-    for (const auto& extension : available_extensions)
-    {
-        required_extensions.erase(extension.extensionName);
-    }
-
-    return required_extensions.empty();
-}
-
-SwapChainSupportDetails SwapChainSupportDetails::query_swap_chain_support(
-    VkPhysicalDevice device, VkSurfaceKHR surface)
-{
-    SwapChainSupportDetails details = {};
-    vkGetPhysicalDeviceSurfaceCapabilitiesKHR(device, surface, &details.capabilities);
-
-    uint32_t format_count = 0;
-    vkGetPhysicalDeviceSurfaceFormatsKHR(device, surface, &format_count, nullptr);
-
-    if (format_count != 0)
-    {
-        details.formats.resize(format_count);
-        vkGetPhysicalDeviceSurfaceFormatsKHR(device, surface, &format_count,
-                                             details.formats.data());
-    }
-
-    uint32_t present_mode_count = 0;
-    vkGetPhysicalDeviceSurfacePresentModesKHR(device, surface, &present_mode_count, nullptr);
-
-    if (present_mode_count != 0)
-    {
-        details.present_modes.resize(present_mode_count);
-        vkGetPhysicalDeviceSurfacePresentModesKHR(device, surface, &present_mode_count,
-                                                  details.present_modes.data());
-    }
-
-    return details;
-}
-
-void Context::create_command_pool()
-{
-    QueueFamilyIndices queue_family_indices =
-        QueueFamilyIndices::find_queue_families(m_physical_device, m_surface);
-
-    VkCommandPoolCreateInfo pool_info = {};
-    pool_info.sType = VK_STRUCTURE_TYPE_COMMAND_POOL_CREATE_INFO;
-    pool_info.flags = VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT;
-    pool_info.queueFamilyIndex = queue_family_indices.graphics_family.value();
-
-    if (vkCreateCommandPool(m_device, &pool_info, nullptr, &m_command_pool) != VK_SUCCESS)
-        throw std::runtime_error("Failed to Create Command Pool!");
-}
-
-QueueFamilyIndices QueueFamilyIndices::find_queue_families(VkPhysicalDevice device,
-                                                           VkSurfaceKHR surface)
-{
-    QueueFamilyIndices indices = {};
-
-    uint32_t queue_family_count = 0;
-    vkGetPhysicalDeviceQueueFamilyProperties(device, &queue_family_count, nullptr);
-
-    std::vector<VkQueueFamilyProperties> queue_families(queue_family_count);
-    vkGetPhysicalDeviceQueueFamilyProperties(device, &queue_family_count, queue_families.data());
-
-    int i = 0;
-    for (const auto& queue_family : queue_families)
-    {
-        VkBool32 present_support = {};
-        vkGetPhysicalDeviceSurfaceSupportKHR(device, i, surface, &present_support);
-
-        if (queue_family.queueFlags & VK_QUEUE_GRAPHICS_BIT)
-            indices.graphics_family = i;
-
-        if (indices.is_complete())
-            break;
-
-        if (present_support)
-            indices.present_family = i;
-
-        i++;
-    }
-
-    return indices;
-}
