--- src/rendering/renderer.cpp
@@ -1,568 +1,0 @@
-#include "renderer.hpp"
-
-#include <algorithm>
-#include <stdexcept>
-#include <fstream>
-
-using namespace molten;
-
-Renderer::Renderer(std::shared_ptr<Context> context) : m_context(context)
-{
-}
-
-void Renderer::init()
-{
-    create_swap_chain();
-    create_image_views();
-    create_graphics_pipeline();
-    create_command_buffers();
-    create_sync_objects();
-}
-
-void molten::Renderer::draw()
-{
-    vkWaitForFences(m_context->m_device, 1, &m_in_flight_fences[current_frame], VK_TRUE, UINT64_MAX);
-
-    uint32_t image_index = UINT64_MAX; // init to high value to avoid confusion
-    VkResult result = vkAcquireNextImageKHR(m_context->m_device, m_swap_chain, UINT64_MAX,
-                                            m_image_available_semaphores[current_frame],
-                                            VK_NULL_HANDLE, &image_index);
-
-    if (result == VK_ERROR_OUT_OF_DATE_KHR)
-    {
-        recreate_swap_chain();
-        return;
-    }
-    else if (result != VK_SUCCESS && result != VK_SUBOPTIMAL_KHR)
-    {
-        throw std::runtime_error("Failed to Acquire Swap Chain Image!");
-    }
-
-    vkResetFences(m_context->m_device, 1, &m_in_flight_fences[current_frame]);
-
-    vkResetCommandBuffer(m_command_buffers[current_frame], 0);
-    record_command_buffer(m_command_buffers[current_frame], image_index);
-
-    VkSubmitInfo submit_info = {};
-    submit_info.sType = VK_STRUCTURE_TYPE_SUBMIT_INFO;
-    VkSemaphore wait_semaphores[] = {m_image_available_semaphores[current_frame]};
-    VkPipelineStageFlags wait_stages[] = {VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT};
-    submit_info.waitSemaphoreCount = 1;
-    submit_info.pWaitSemaphores = wait_semaphores;
-    submit_info.pWaitDstStageMask = wait_stages;
-    submit_info.commandBufferCount = 1;
-    submit_info.pCommandBuffers = &m_command_buffers[current_frame];
-    VkSemaphore signal_semaphores[] = {m_render_finished_semaphores[current_frame]};
-    submit_info.signalSemaphoreCount = 1;
-    submit_info.pSignalSemaphores = signal_semaphores;
-
-    if (vkQueueSubmit(m_context->m_graphics_queue, 1, &submit_info,
-                      m_in_flight_fences[current_frame]) !=
-        VK_SUCCESS)
-        throw std::runtime_error("Failed to Submit Draw Command Buffer!");
-
-    VkPresentInfoKHR present_info = {};
-    present_info.sType = VK_STRUCTURE_TYPE_PRESENT_INFO_KHR;
-    present_info.waitSemaphoreCount = 1;
-    present_info.pWaitSemaphores = signal_semaphores;
-
-    VkSwapchainKHR swap_chains[] = {m_swap_chain};
-    present_info.swapchainCount = 1;
-    present_info.pSwapchains = swap_chains;
-    present_info.pImageIndices = &image_index;
-    present_info.pResults = nullptr;
-
-    result = vkQueuePresentKHR(m_context->m_present_queue, &present_info);
-
-    if (result == VK_ERROR_OUT_OF_DATE_KHR || result == VK_SUBOPTIMAL_KHR ||
-        m_context->m_framebuffer_resized)
-        recreate_swap_chain();
-    else if (result != VK_SUCCESS)
-        throw std::runtime_error("Failed to Present Swap Chain Image!");
-
-    current_frame = (current_frame + 1) % MAX_FRAMES_IN_FLIGHT;
-}
-
-void molten::Renderer::cleanup()
-{
-    cleanup_swap_chain();
-
-    vkDestroyPipeline(m_context->m_device, m_graphics_pipeline, nullptr);
-    vkDestroyPipelineLayout(m_context->m_device, m_pipeline_layout, nullptr);
-
-    for (size_t i = 0; i < MAX_FRAMES_IN_FLIGHT; i++)
-    {
-        vkDestroySemaphore(m_context->m_device, m_render_finished_semaphores[i], nullptr);
-        vkDestroySemaphore(m_context->m_device, m_image_available_semaphores[i], nullptr);
-        vkDestroyFence(m_context->m_device, m_in_flight_fences[i], nullptr);
-    }
-}
-
-VkSurfaceFormatKHR Renderer::choose_swap_surface_format(
-    const std::vector<VkSurfaceFormatKHR>& available_formats)
-{
-    for (const auto& available_format : available_formats)
-    {
-        if (available_format.format == VK_FORMAT_B8G8R8A8_SRGB &&
-            available_format.colorSpace == VK_COLOR_SPACE_SRGB_NONLINEAR_KHR)
-            return available_format;
-    }
-    return available_formats[0];
-}
-
-VkPresentModeKHR Renderer::choose_swap_present_mode(
-    const std::vector<VkPresentModeKHR>& available_present_modes)
-{
-    for (const auto& available_present_mode : available_present_modes)
-    {
-        if (available_present_mode == VK_PRESENT_MODE_MAILBOX_KHR)
-            return available_present_mode;
-    }
-    return VK_PRESENT_MODE_FIFO_KHR;
-}
-
-VkExtent2D Renderer::choose_swap_extent(const VkSurfaceCapabilitiesKHR& capabilities)
-{
-    if (capabilities.currentExtent.width != (std::numeric_limits<uint32_t>::max)())
-    {
-        return capabilities.currentExtent;
-    }
-    else
-    {
-        int width = -1, height = -1;
-        glfwGetFramebufferSize(m_context->m_window, &width, &height);
-
-        VkExtent2D actual_extent = {static_cast<uint32_t>(width), static_cast<uint32_t>(height)};
-
-        actual_extent.width = std::clamp(actual_extent.width, capabilities.minImageExtent.width,
-                                         capabilities.maxImageExtent.width);
-        actual_extent.height = std::clamp(actual_extent.height, capabilities.minImageExtent.height,
-                                          capabilities.maxImageExtent.height);
-        return actual_extent;
-    }
-}
-
-void Renderer::create_swap_chain()
-{
-    SwapChainSupportDetails swap_chain_support =
-        SwapChainSupportDetails::query_swap_chain_support(m_context->m_physical_device, m_context->m_surface);
-
-    VkSurfaceFormatKHR surface_format = choose_swap_surface_format(swap_chain_support.formats);
-    VkPresentModeKHR present_mode = choose_swap_present_mode(swap_chain_support.present_modes);
-    VkExtent2D extent = choose_swap_extent(swap_chain_support.capabilities);
-
-    uint32_t image_count = swap_chain_support.capabilities.minImageCount + 1;
-    if (swap_chain_support.capabilities.maxImageCount > 0 &&
-        image_count > swap_chain_support.capabilities.maxImageCount)
-        image_count = swap_chain_support.capabilities.maxImageCount;
-
-    VkSwapchainCreateInfoKHR create_info = {};
-    create_info.sType = VK_STRUCTURE_TYPE_SWAPCHAIN_CREATE_INFO_KHR;
-    create_info.surface = m_context->m_surface;
-    create_info.minImageCount = image_count;
-    create_info.imageFormat = surface_format.format;
-    create_info.imageColorSpace = surface_format.colorSpace;
-    create_info.imageExtent = extent;
-    create_info.imageArrayLayers = 1;
-    create_info.imageUsage = VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT;
-
-    QueueFamilyIndices indices =
-        QueueFamilyIndices::find_queue_families(m_context->m_physical_device, m_context->m_surface);
-    uint32_t queue_family_indices[] = {indices.graphics_family.value(),
-                                       indices.present_family.value()};
-
-    if (indices.graphics_family != indices.present_family)
-    {
-        create_info.imageSharingMode = VK_SHARING_MODE_CONCURRENT;
-        create_info.queueFamilyIndexCount = 2;
-        create_info.pQueueFamilyIndices = queue_family_indices;
-    }
-    else
-    {
-        create_info.imageSharingMode = VK_SHARING_MODE_EXCLUSIVE;
-        create_info.queueFamilyIndexCount = 0;
-        create_info.pQueueFamilyIndices = nullptr;
-    }
-    create_info.preTransform = swap_chain_support.capabilities.currentTransform;
-    create_info.compositeAlpha = VK_COMPOSITE_ALPHA_OPAQUE_BIT_KHR;
-    create_info.presentMode = present_mode;
-    create_info.clipped = VK_TRUE;
-    create_info.oldSwapchain = VK_NULL_HANDLE;
-
-    if (vkCreateSwapchainKHR(m_context->m_device, &create_info, nullptr, &m_swap_chain) !=
-        VK_SUCCESS)
-        throw std::runtime_error("Failed to Create Swap Chain!");
-
-    vkGetSwapchainImagesKHR(m_context->m_device, m_swap_chain, &image_count, nullptr);
-    m_swap_chain_images.resize(image_count);
-    vkGetSwapchainImagesKHR(m_context->m_device, m_swap_chain, &image_count,
-                            m_swap_chain_images.data());
-
-    m_swap_chain_image_format = surface_format.format;
-    m_swap_chain_extent = extent;
-}
-
-void Renderer::recreate_swap_chain()
-{
-    int width = 0, height = 0;
-    glfwGetFramebufferSize(m_context->m_window, &width, &height);
-    while (width == 0 || height == 0)
-    {
-        glfwGetFramebufferSize(m_context->m_window, &width, &height);
-        glfwWaitEvents();
-    }
-
-    vkDeviceWaitIdle(m_context->m_device);
-
-    cleanup_swap_chain();
-
-    create_swap_chain();
-    create_image_views();
-}
-
-void Renderer::create_image_views()
-{
-    m_swap_chain_image_views.resize(m_swap_chain_images.size());
-    for (size_t i = 0; i < m_swap_chain_images.size(); i++)
-    {
-        VkImageViewCreateInfo create_info = {};
-        create_info.sType = VK_STRUCTURE_TYPE_IMAGE_VIEW_CREATE_INFO;
-        create_info.image = m_swap_chain_images[i];
-        create_info.viewType = VK_IMAGE_VIEW_TYPE_2D;
-        create_info.format = m_swap_chain_image_format;
-        create_info.components.r = VK_COMPONENT_SWIZZLE_IDENTITY;
-        create_info.components.g = VK_COMPONENT_SWIZZLE_IDENTITY;
-        create_info.components.b = VK_COMPONENT_SWIZZLE_IDENTITY;
-        create_info.components.a = VK_COMPONENT_SWIZZLE_IDENTITY;
-        create_info.subresourceRange.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-        create_info.subresourceRange.baseMipLevel = 0;
-        create_info.subresourceRange.levelCount = 1;
-        create_info.subresourceRange.baseArrayLayer = 0;
-        create_info.subresourceRange.layerCount = 1;
-        if (vkCreateImageView(m_context->m_device, &create_info, nullptr,
-                              &m_swap_chain_image_views[i]) !=
-            VK_SUCCESS)
-            throw std::runtime_error("Failed to Create Image Views");
-    }
-}
-
-void molten::Renderer::cleanup_swap_chain()
-{
-    /*for (auto framebuffer : m_swap_chain_frame_buffers)
-    {
-        vkDestroyFramebuffer(m_device, framebuffer, nullptr);
-    }*/
-
-    for (auto image_view : m_swap_chain_image_views)
-    {
-        vkDestroyImageView(m_context->m_device, image_view, nullptr);
-    }
-
-    vkDestroySwapchainKHR(m_context->m_device, m_swap_chain, nullptr);
-}
-
-void Renderer::create_graphics_pipeline()
-{
-    auto vert_shader_code = read_file("shaders/vert.spv");
-    auto frag_shader_code = read_file("shaders/frag.spv");
-
-    VkShaderModule vert_shader_module = create_shader_module(vert_shader_code);
-    VkShaderModule frag_shader_module = create_shader_module(frag_shader_code);
-
-    VkPipelineShaderStageCreateInfo vert_shader_stage_info = {};
-    vert_shader_stage_info.sType = VK_STRUCTURE_TYPE_PIPELINE_SHADER_STAGE_CREATE_INFO;
-    vert_shader_stage_info.stage = VK_SHADER_STAGE_VERTEX_BIT;
-    vert_shader_stage_info.module = vert_shader_module;
-    vert_shader_stage_info.pName = "main";
-
-    VkPipelineShaderStageCreateInfo frag_shader_stage_info = {};
-    frag_shader_stage_info.sType = VK_STRUCTURE_TYPE_PIPELINE_SHADER_STAGE_CREATE_INFO;
-    frag_shader_stage_info.stage = VK_SHADER_STAGE_FRAGMENT_BIT;
-    frag_shader_stage_info.module = frag_shader_module;
-    frag_shader_stage_info.pName = "main";
-
-    VkPipelineShaderStageCreateInfo shader_stages[] = {vert_shader_stage_info,
-                                                       frag_shader_stage_info};
-
-    std::vector<VkDynamicState> dynamic_states = {VK_DYNAMIC_STATE_VIEWPORT,
-                                                  VK_DYNAMIC_STATE_SCISSOR};
-
-    VkPipelineDynamicStateCreateInfo dynamic_state = {};
-    dynamic_state.sType = VK_STRUCTURE_TYPE_PIPELINE_DYNAMIC_STATE_CREATE_INFO;
-    dynamic_state.dynamicStateCount = static_cast<uint32_t>(dynamic_states.size());
-    dynamic_state.pDynamicStates = dynamic_states.data();
-
-    VkPipelineVertexInputStateCreateInfo vertex_input_info = {};
-    vertex_input_info.sType = VK_STRUCTURE_TYPE_PIPELINE_VERTEX_INPUT_STATE_CREATE_INFO;
-    vertex_input_info.vertexBindingDescriptionCount = 0;
-    vertex_input_info.pVertexBindingDescriptions = nullptr;
-    vertex_input_info.vertexAttributeDescriptionCount = 0;
-    vertex_input_info.pVertexAttributeDescriptions = nullptr;
-
-    VkPipelineInputAssemblyStateCreateInfo input_assembly = {};
-    input_assembly.sType = VK_STRUCTURE_TYPE_PIPELINE_INPUT_ASSEMBLY_STATE_CREATE_INFO;
-    input_assembly.topology = VK_PRIMITIVE_TOPOLOGY_TRIANGLE_LIST;
-    input_assembly.primitiveRestartEnable = VK_FALSE;
-
-    VkViewport viewport = {};
-    viewport.x = 0.0f;
-    viewport.y = 0.0f;
-    viewport.width = (float)m_swap_chain_extent.width;
-    viewport.height = (float)m_swap_chain_extent.height;
-    viewport.minDepth = 0.0f;
-    viewport.maxDepth = 1.0f;
-    VkRect2D scissor = {};
-    scissor.offset = {0, 0};
-    scissor.extent = m_swap_chain_extent;
-
-    VkPipelineViewportStateCreateInfo viewport_state = {};
-    viewport_state.sType = VK_STRUCTURE_TYPE_PIPELINE_VIEWPORT_STATE_CREATE_INFO;
-    viewport_state.viewportCount = 1;
-    viewport_state.pViewports = &viewport;
-    viewport_state.scissorCount = 1;
-    viewport_state.pScissors = &scissor;
-
-    VkPipelineRasterizationStateCreateInfo rasterizer = {};
-    rasterizer.sType = VK_STRUCTURE_TYPE_PIPELINE_RASTERIZATION_STATE_CREATE_INFO;
-    rasterizer.depthClampEnable = VK_FALSE;
-    rasterizer.rasterizerDiscardEnable = VK_FALSE;
-    rasterizer.polygonMode = VK_POLYGON_MODE_FILL;
-    rasterizer.lineWidth = 1.0f;
-    rasterizer.cullMode = VK_CULL_MODE_BACK_BIT;
-    rasterizer.frontFace = VK_FRONT_FACE_CLOCKWISE;
-    rasterizer.depthBiasEnable = VK_FALSE;
-    rasterizer.depthBiasConstantFactor = 0.0f;
-    rasterizer.depthBiasClamp = 0.0f;
-    rasterizer.depthBiasSlopeFactor = 0.0f;
-
-    VkPipelineMultisampleStateCreateInfo multisampling = {};
-    multisampling.sType = VK_STRUCTURE_TYPE_PIPELINE_MULTISAMPLE_STATE_CREATE_INFO;
-    multisampling.sampleShadingEnable = VK_FALSE;
-    multisampling.rasterizationSamples = VK_SAMPLE_COUNT_1_BIT;
-    multisampling.minSampleShading = 1.0f;
-    multisampling.pSampleMask = nullptr;
-    multisampling.alphaToCoverageEnable = VK_FALSE;
-    multisampling.alphaToOneEnable = VK_FALSE;
-
-    VkPipelineColorBlendAttachmentState color_blend_attachment = {};
-    color_blend_attachment.colorWriteMask = VK_COLOR_COMPONENT_R_BIT | VK_COLOR_COMPONENT_G_BIT |
-                                            VK_COLOR_COMPONENT_B_BIT | VK_COLOR_COMPONENT_A_BIT;
-    color_blend_attachment.blendEnable = VK_FALSE;
-    color_blend_attachment.srcColorBlendFactor = VK_BLEND_FACTOR_SRC_ALPHA;
-    color_blend_attachment.dstColorBlendFactor = VK_BLEND_FACTOR_ONE_MINUS_SRC_ALPHA;
-    color_blend_attachment.colorBlendOp = VK_BLEND_OP_ADD;
-    color_blend_attachment.srcAlphaBlendFactor = VK_BLEND_FACTOR_ONE;
-    color_blend_attachment.dstAlphaBlendFactor = VK_BLEND_FACTOR_ZERO;
-    color_blend_attachment.alphaBlendOp = VK_BLEND_OP_ADD;
-
-    VkPipelineColorBlendStateCreateInfo color_blending = {};
-    color_blending.sType = VK_STRUCTURE_TYPE_PIPELINE_COLOR_BLEND_STATE_CREATE_INFO;
-    color_blending.logicOpEnable = VK_FALSE;
-    color_blending.logicOp = VK_LOGIC_OP_COPY;
-    color_blending.attachmentCount = 1;
-    color_blending.pAttachments = &color_blend_attachment;
-    color_blending.blendConstants[0] = 0.0f;
-    color_blending.blendConstants[1] = 0.0f;
-    color_blending.blendConstants[2] = 0.0f;
-    color_blending.blendConstants[3] = 0.0f;
-
-    VkPipelineLayoutCreateInfo pipeline_layout_info = {};
-    pipeline_layout_info.sType = VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO;
-    pipeline_layout_info.setLayoutCount = 0;
-    pipeline_layout_info.pSetLayouts = nullptr;
-    pipeline_layout_info.pushConstantRangeCount = 0;
-    pipeline_layout_info.pPushConstantRanges = nullptr;
-
-    if (vkCreatePipelineLayout(m_context->m_device, &pipeline_layout_info, nullptr, &m_pipeline_layout) !=
-        VK_SUCCESS)
-        throw std::runtime_error("Failed to Create Pipeline Layout!");
-
-    VkPipelineRenderingCreateInfoKHR pipeline_rendering_info = {};
-    pipeline_rendering_info.sType = VK_STRUCTURE_TYPE_PIPELINE_RENDERING_CREATE_INFO_KHR;
-    pipeline_rendering_info.colorAttachmentCount = 1;
-    pipeline_rendering_info.pColorAttachmentFormats = &m_swap_chain_image_format;
-
-    VkGraphicsPipelineCreateInfo pipeline_info = {};
-    pipeline_info.sType = VK_STRUCTURE_TYPE_GRAPHICS_PIPELINE_CREATE_INFO;
-    pipeline_info.stageCount = 2;
-    pipeline_info.pStages = shader_stages;
-    pipeline_info.pVertexInputState = &vertex_input_info;
-    pipeline_info.pInputAssemblyState = &input_assembly;
-    pipeline_info.pViewportState = &viewport_state;
-    pipeline_info.pRasterizationState = &rasterizer;
-    pipeline_info.pMultisampleState = &multisampling;
-    pipeline_info.pDepthStencilState = nullptr;
-    pipeline_info.pColorBlendState = &color_blending;
-    pipeline_info.pDynamicState = &dynamic_state;
-    pipeline_info.layout = m_pipeline_layout;
-    pipeline_info.renderPass = nullptr /*m_render_pass*/;
-    pipeline_info.pNext = &pipeline_rendering_info;
-    pipeline_info.subpass = 0;
-    pipeline_info.basePipelineHandle = VK_NULL_HANDLE;
-    pipeline_info.basePipelineIndex = -1;
-
-    if (vkCreateGraphicsPipelines(m_context->m_device, VK_NULL_HANDLE, 1, &pipeline_info, nullptr,
-                                  &m_graphics_pipeline) != VK_SUCCESS)
-        throw std::runtime_error("Failed to Create Graphics Pipeline!");
-
-    vkDestroyShaderModule(m_context->m_device, frag_shader_module, nullptr);
-    vkDestroyShaderModule(m_context->m_device, vert_shader_module, nullptr);
-}
-
-std::vector<char> Renderer::read_file(const std::string& filename)
-{
-    std::ifstream file(filename, std::ios::ate | std::ios::binary);
-
-    if (!file.is_open())
-        throw std::runtime_error("Failed to Open File!");
-
-    size_t file_size = (size_t)file.tellg();
-    printf("File size of %s is %zd \n", filename.c_str(), file_size);
-    std::vector<char> buffer(file_size);
-
-    file.seekg(0);
-    file.read(buffer.data(), file_size);
-
-    file.close();
-    return buffer;
-}
-
-VkShaderModule Renderer::create_shader_module(const std::vector<char>& code)
-{
-    VkShaderModuleCreateInfo create_info = {};
-    create_info.sType = VK_STRUCTURE_TYPE_SHADER_MODULE_CREATE_INFO;
-    create_info.codeSize = code.size();
-    create_info.pCode = reinterpret_cast<const uint32_t*>(code.data());
-
-    VkShaderModule shader_module = {};
-    if (vkCreateShaderModule(m_context->m_device, &create_info, nullptr, &shader_module) !=
-        VK_SUCCESS)
-        throw std::runtime_error("Failed to Create Shader Module!");
-
-    return shader_module;
-}
-
-void Renderer::create_command_buffers()
-{
-    m_command_buffers.resize(MAX_FRAMES_IN_FLIGHT);
-
-    VkCommandBufferAllocateInfo alloc_info = {};
-    alloc_info.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_ALLOCATE_INFO;
-    alloc_info.commandPool = m_context->m_command_pool;
-    alloc_info.level = VK_COMMAND_BUFFER_LEVEL_PRIMARY;
-    alloc_info.commandBufferCount = (uint32_t)m_command_buffers.size();
-
-    if (vkAllocateCommandBuffers(m_context->m_device, &alloc_info, m_command_buffers.data()) !=
-        VK_SUCCESS)
-        throw std::runtime_error("Failed to Allocate Command Buffers!");
-}
-
-void Renderer::record_command_buffer(VkCommandBuffer command_buffer, uint32_t image_index)
-{
-    VkCommandBufferBeginInfo begin_info = {};
-    begin_info.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_BEGIN_INFO;
-    begin_info.flags = 0;
-    begin_info.pInheritanceInfo = nullptr;
-
-    if (vkBeginCommandBuffer(command_buffer, &begin_info) != VK_SUCCESS)
-        throw std::runtime_error("Failed to Begin Recording Command Buffer!");
-
-    VkClearValue clear_color = {{{0.0f, 0.0f, 0.0f, 1.0f}}};
-
-    VkRenderingAttachmentInfoKHR color_attachment_info = {};
-    color_attachment_info.sType = VK_STRUCTURE_TYPE_RENDERING_ATTACHMENT_INFO_KHR;
-    color_attachment_info.imageView = m_swap_chain_image_views[image_index];
-    color_attachment_info.imageLayout = VK_IMAGE_LAYOUT_ATTACHMENT_OPTIMAL_KHR;
-    color_attachment_info.loadOp = VK_ATTACHMENT_LOAD_OP_CLEAR;
-    color_attachment_info.storeOp = VK_ATTACHMENT_STORE_OP_STORE;
-    color_attachment_info.clearValue = clear_color;
-    VkRenderingInfoKHR render_info = {};
-    render_info.sType = VK_STRUCTURE_TYPE_RENDERING_INFO_KHR;
-    render_info.renderArea.offset = {0, 0};
-    render_info.renderArea.extent = m_swap_chain_extent;
-    render_info.layerCount = 1;
-    render_info.colorAttachmentCount = 1;
-    render_info.pColorAttachments = &color_attachment_info;
-
-    VkImageMemoryBarrier image_memory_barrier = {};
-    image_memory_barrier.sType = VK_STRUCTURE_TYPE_IMAGE_MEMORY_BARRIER;
-    image_memory_barrier.srcAccessMask = VK_ACCESS_COLOR_ATTACHMENT_WRITE_BIT;
-    image_memory_barrier.oldLayout = VK_IMAGE_LAYOUT_UNDEFINED;
-    image_memory_barrier.newLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-    image_memory_barrier.image = m_swap_chain_images[image_index];
-    VkImageSubresourceRange subresource_range = {};
-    subresource_range.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-    subresource_range.baseMipLevel = 0;
-    subresource_range.levelCount = 1;
-    subresource_range.baseArrayLayer = 0;
-    subresource_range.layerCount = 1;
-    image_memory_barrier.subresourceRange = subresource_range;
-
-    vkCmdPipelineBarrier(command_buffer, VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT,
-                         VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT, 0, 0, nullptr, 0, nullptr,
-                         1, &image_memory_barrier);
-
-    m_context->BeginRendering(command_buffer, &render_info); // vkCmdBeginRenderingKHR
-
-    vkCmdBindPipeline(command_buffer, VK_PIPELINE_BIND_POINT_GRAPHICS, m_graphics_pipeline);
-
-    VkViewport viewport = {};
-    viewport.x = 0.0f;
-    viewport.y = 0.0f;
-    viewport.width = static_cast<float>(m_swap_chain_extent.width);
-    viewport.height = static_cast<float>(m_swap_chain_extent.height);
-    viewport.minDepth = 0.0f;
-    viewport.maxDepth = 1.0f;
-    vkCmdSetViewport(command_buffer, 0, 1, &viewport);
-
-    VkRect2D scissor = {};
-    scissor.offset = {0, 0};
-    scissor.extent = m_swap_chain_extent;
-    vkCmdSetScissor(command_buffer, 0, 1, &scissor);
-
-    vkCmdDraw(command_buffer, 3, 1, 0, 0);
-
-    m_context->EndRendering(command_buffer); // vkCmdEndRenderingKHR
-
-    VkImageMemoryBarrier im_mem_barrier = {};
-    im_mem_barrier.sType = VK_STRUCTURE_TYPE_IMAGE_MEMORY_BARRIER;
-    im_mem_barrier.srcAccessMask = VK_ACCESS_COLOR_ATTACHMENT_WRITE_BIT;
-    im_mem_barrier.oldLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-    im_mem_barrier.newLayout = VK_IMAGE_LAYOUT_PRESENT_SRC_KHR;
-    im_mem_barrier.image = m_swap_chain_images[image_index];
-    im_mem_barrier.subresourceRange = subresource_range;
-
-    vkCmdPipelineBarrier(command_buffer, VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT,
-                         VK_PIPELINE_STAGE_BOTTOM_OF_PIPE_BIT, 0, 0, nullptr, 0, nullptr, 1,
-                         &im_mem_barrier);
-
-    if (vkEndCommandBuffer(command_buffer) != VK_SUCCESS)
-        throw std::runtime_error("Failed to Record Command Buffer!");
-}
-
-void Renderer::create_sync_objects()
-{
-    m_image_available_semaphores.resize(MAX_FRAMES_IN_FLIGHT);
-    m_render_finished_semaphores.resize(MAX_FRAMES_IN_FLIGHT);
-    m_in_flight_fences.resize(MAX_FRAMES_IN_FLIGHT);
-
-    VkSemaphoreCreateInfo semaphore_info = {};
-    semaphore_info.sType = VK_STRUCTURE_TYPE_SEMAPHORE_CREATE_INFO;
-
-    VkFenceCreateInfo fence_info = {};
-    fence_info.sType = VK_STRUCTURE_TYPE_FENCE_CREATE_INFO;
-    fence_info.flags = VK_FENCE_CREATE_SIGNALED_BIT;
-
-    for (size_t i = 0; i < MAX_FRAMES_IN_FLIGHT; i++)
-    {
-        if (vkCreateSemaphore(m_context->m_device, &semaphore_info, nullptr,
-                              &m_image_available_semaphores[i]) != VK_SUCCESS ||
-            vkCreateSemaphore(m_context->m_device, &semaphore_info, nullptr,
-                              &m_render_finished_semaphores[i]) != VK_SUCCESS ||
-            vkCreateFence(m_context->m_device, &fence_info, nullptr, &m_in_flight_fences[i]) !=
-                VK_SUCCESS)
-        {
-            throw std::runtime_error("Failed to Create Semaphores and Fences!");
-        }
-    }
-}
